Don't retry individual SQL transaction commands
If anything in a transaction fails, the transaction should be retried as
a whole, just retrying some individual command makes no sense and breaks
transaction semantics. For example, if a COMMIT fails in MySQL, it
automatically rolls back the transaction and if you issue another
COMMIT, you just commit an empty transaction.

Remove the test case for commit as is basically tested that it showed
the broken behavior described above.