Make overdue sync HA-aware
Syncing the overdue flag to the host_state and service_state tables was
not HA-aware and assumed to have a consistent copy of all serivces/hosts
with overdue set stored in Redis. However, if a second Icinga DB
instance was active, this set might have become outdated.

This commit fixes this by making the overdue sync HA-aware, i.e. pausing
it if the current instance is inactive. Additionally, it fetches the set
of overdue services from MySQL and syncs it back to Redis when becoming
active.