configsync: use HSCAN instead of HKEYS
Redis documentation states that it is preferable to use *SCAN instead of
*KEYS (https://redis.io/commands/keys):

  Warning: consider KEYS as a command that should only be used in
  production environments with extreme care. It may ruin performance
  when it is executed against large databases. This command is intended
  for debugging and special operations, such as changing your keyspace
  layout.  Don't use KEYS in your regular application code. If you're
  looking for a way to find keys in a subset of your keyspace, consider
  using SCAN or sets.

While the documentation is for the KEYS command and not HKEYS, the same
warning should apply here to as the hashes we are iterating over can get
quite large. Therefore use the HSCAN command that exists as an
alternative.