Use the introduced Valuer everywhere truncation is needed
The Valuer introduced from the previous commit is used to perform the truncation on the columns with text data type. In other words,
the type of the field is changed from string to Txtcol or MediumTxtcol or Perfcol (based on whether it is a text or medium text
or performance data respectively) where the truncation is required.