Add option to send log messages to systemd journal
`output` field is added to `logging` section in `config.yml` file. If it is set to `systemd` then logs are sent to systemd journal.
By default it is set to `console`. `journal.go` is added to `logging` package in which the functionality to send the messages to
systemd journal is written.