Remove `output` from `logging` in config.yml
Get whether the log outputs need to be written to stdErr or sent to systemd journal using daemon.SdNotify(false, daemon.SdNotifyReady) and os.LookupEnv("NOTIFY_SOCKET").
And in icingadb.service when `Type` is set to `simple` then the output will be `console` and when it is set to `notify` the output will be `systemd-journa`.

In this commit there are also changes to the code document.