Add unmarshaller to config.go and change logging level to type zapcore.Level
Unmarshaller is added to config.go to read sub fields which are set to default, if the main field is not present in config.yml file.
For example logging has all optional fields and hence  may not be included sometimes in config.yml file, in which case the default logging level is set to debug.