Make sure to set default values for config options
If no option is specified in the configuration,
UnmarshalYAML is not called for the Options structs and
therefore no default values are set.

In the config structs we now check whether Options is nil,
which means that no options are set in the configuration.
Then we call unmarshal ourselves to trigger the
Options.UnmarshalYAML call to set the default values.

Alternatively, we could also call defaults.Set() here, but then,
when changing the UnmarshalYAML functions,
we would also have to check here whether changes need to be made.