Use own connections for heartbeat and HA
Use dedicated connections for heartbeat and HA to ensure that
heartbeats are always processed and the instance table is updated.
Otherwise, the connections can be too busy due to the
synchronization of configuration, status, history, etc.,
which can lead to handover / takeover loops because the heartbeat
is not read while HA gets stuck when updating the instance table.