List required RDBMS versions and increase PostgreSQL minimum
MySQL minimum versions are based on what was used in Icinga DB 1.0.0 RC1.

The required PostgreSQL version is bumped to at least 9.6 as this allows to
label functions as `PARALLEL RESTRICTED` like it's done by the SLA reporting.
This or newer versions are available for both CentOS/RHEL 8 and SLES 12 which
were the primary constraints for the previous version requirement.

For CentOS/RHEL, don't specify some additional paths and package names with a
specific version, as any newer version is also fine and there's a choice of
multiple versions. Also, we don't guide the user through actually installing
the PostgreSQL server, so they should already be familiar with the differing
names if their distribution reuqires them.

[1] https://github.com/Icinga/icingadb/blob/v1.0.0-rc1/.github/workflows/go.yml