Fix tests for new icinga2 master
A recent change in icinga2 [1] prevents modifying some object attributes via
the API, most notably those that influence the name of the object. Previously,
the tests would also include these attributes in the POST requests that modify
the objects (but with the same value as before), something that is now
rejected.  Therefore, this commits allows to label attributes as "nomodify" so
that they won't be sent in this case.

[1]: https://github.com/Icinga/icinga2/pull/9407