Make integration tests work with Go 1.20.6
As a security fix, Go implemented stricter checks for the Host header in the
http package. This breaks the Go Docker client library [1] resulting in "http:
invalid Host header" errors in our integration tests.

A fix [2] was merged to the 24.0 branch, but is not yet included in a release.
Therefore, this commit updates the dependency to a specific commit
(`go get -u github.com/docker/docker@24.0`). Once it's included in a release,
this can be switched back to using a release version.

[1] https://github.com/moby/moby/issues/45935
[2] https://github.com/moby/moby/pull/45942