GitHub Actions: use actions/setup-go@v4 and go-version: 1.x
There was a mix of diferent action and Go versions specified throughout the
GitHub Actions workflow configs. In general, we want to be compatible with the
latest Go 1.x version, so just specify that instead of bumping versions
everywhere when necessary.

At the same time, bump the actions/setup-go to v4, the latest currently
available version. This prefers versions already cached on the runner and
enables automatic caching for dependencies.