GitHub Actions: remove version check
This check turned out to be not very helpful. Typically, you will first
increase the version number in the source, make it a PR, review it, merge it
and only aftewards, tag the version. In this PR, the check is always
complaining if you did not forget to do something.