History Sync: Use single, distinct DB connection
The only database tables with foreign key constraints are found in the
history context. By default - and for good reasons - a DB internally
consists of a cluster of several database connections. However, if a
multi-master database server setup is used, it can - and will - happen
that this DB contains sessions to different servers and related queries
will be submitted in the wrong order, at least in the eye of some
database servers.

To prevent this problem, an extra DB with only one connection is used
for the history sync. First, this required some refactoring on
icingadb.DB, allowing to clone, copy or re-create a DB.