types: Add UnixNano based on UnixMilli
As Icinga/icinga-notifications#61 required a more precise time type, the
UnixMilli type was copied to UnixNano, representing nanoseconds.

While implementing the relevant changes, some changes were backported to
UnixMilli. For example, the util.FromUnixMilli function is no longer
used and was marked as deprecated as there is now time.UnixMilli in the
standard library.