HA State: ensure not this and another instance can be responsible
In theory, this should not happen. This assumption is based on the trust
in the database transaction performing the HA realization logic.
However, one debugged log let one assume that this happened anyway.

This change mostly signals an error while also explicitly giving up the
HA state. Doing so should at least alarm a person reading the logs.