doc/03-Configuration.md: Environment Variables
New ways to configure Icinga DB were introduced in [icingadb-831] and
[igl-113]. The first change allows configuring Icinga DB using
environment variables instead of or next to the YAML configuration file.
In addition, the second change allows setting certificates and keys as
PEM-encoded strings next to referencing files. This was now documented.

As a structural change, the order of the Database and Redis sections
were changed to reflect the order in the example configuration file.

Some words about the Logging Components were written, as the
documentation previously that they exist.

[icingadb-831]: https://github.com/Icinga/icingadb/pull/831
[igl-113]: https://github.com/Icinga/icinga-go-library/pull/113