history: YAML Decoding For RetentionOptions
For YAML decoding, the yaml.InterfaceUnmarshaler interface needs to be
implemented. To not mix value and pointer receiving methods for
RetentionOptions, the signature of Validate was updated.

It is important to set the internally type to map[string]uint16 instead
of RetentionOptions, as otherwise the method would be called in an
endless loop until the stack is full.