Containerfile: allow running the binary without absolute path
This sets the PATH environment variable so that the binary installed to the
container image can be found by just its name. This makes it nicer to manually
provide a command by the container, so for example, now "docker run --rm -it
icinga/icingadb:dev icingadb --version" works. The binary is still installed
under the same path, so it can still be invoked using the same absolute path as
before.