ha: More logging in OnRetryableError in HA.realize
The current state of the OnRetryableError function for the inner
function in HA.realize was designed quite silent. In this changed
version reoccurring errors will be logged and the log level after the
first three attempts was raised to warning. Additionally, the current
attempt and the total duration will be logged as well.