icingadb: Abort with a message instead of a stack trace
When the icingadb application was aborted with a fatal log event, the
user received an impressive stack trace. Unfortunately, the relevant
information is not directly visible for the untrained eye.

This change now prints a normal log event at FATAL level with an error
attached. If an error implements the fmt.Formatter, the zap logging
framework adds both an "error" and an "errorVerbose" field. So we end up
with a short error description, the error string concatenated of all
wrapped errors, and a stack trace.

Since error and errorVerbose are both fields, it is possible to filter
them for the systemd journald logger at a later point.

> 2025-04-15T16:43:36.974+0200    FATAL   icingadb        Can't connect to database       {"error": "can't connect to database: dial tcp [::1]:5432: connect: connection refused", "errorVerbose": "dial tcp [::1]:5432: connect: connection refused\ncan't connect to database\ngithub.com/icinga/icinga-go-library/database.RetryConnector.Connect\n\tgithub.com/icinga/icinga-go-library@v0.6.3/database/driver.go:48\ndatabase/sql.(*DB).conn\n\tdatabase/sql/sql.go:1431\ndatabase/sql.(*DB).PingContext.func1\n\tdatabase/sql/sql.go:900\ndatabase/sql.(*DB).retry\n\tdatabase/sql/sql.go:1576\ndatabase/sql.(*DB).PingContext\n\tdatabase/sql/sql.go:899\ndatabase/sql.(*DB).Ping\n\tdatabase/sql/sql.go:917\nmain.run\n\tgithub.com/icinga/icingadb/cmd/icingadb/main.go:64\nmain.main\n\tgithub.com/icinga/icingadb/cmd/icingadb/main.go:37\nruntime.main\n\truntime/proc.go:283\nruntime.goexit\n\truntime/asm_amd64.s:1700"}