doc/01-About.md: Big Picture
Inspired by the "Big Picture" section in the Icinga Notifications
documentation[^0], such a section was introduced for Icinga DB as well.
It aims to describe the essential components in an Icinga DB setup along
the architecture diagram for novice users. This hopefully helps new
users to understand what Icinga DB actually is and how it interconnects.

[^0]: https://icinga.com/docs/icinga-notifications/latest/#big-picture