icingadb-migrate: Mitigate invalid FROM_UNIXTIME arguments
Ensure that no invalid FROM_UNIXTIME SQL function calls result in no
data being marked as migration-worthy in icingadb-migrate.

MariaDB changed the behavior of the FROM_UNIXTIME function[^0] in
version 11.7, resulting in FROM_UNIXTIME(0) now returning NULL instead
of "1970-01-01 00:00:00". The icingadb-migrate utility uses this SQL
function within the "computeIdRange" function for "ido.from" and
"ido.to".  Since "ido.from" defaults to 0, a NULL comparison occurs in
the WHERE clause.  This results in an empty result set and no data is
marked as migration-worthy.

This fix wraps the user-supplied or default FROM_UNIXTIME value in a
COALESCE with either an infimum or a supremum. If the first
FROM_UNIXTIME returns NULL, the second returns a valid border.

Fixes #974.

[^0]: https://mariadb.com/kb/en/from_unixtime/