docs: PostgreSQL User for Schema Upgrades
Explicitly state that the "icingadb" PostgreSQL database user must be
used for schema upgrades.

The PostgreSQL database user applying a schema upgrade is the table
owner of newly create tables. When applying an Icinga DB schema upgrade
not as the "icingadb" user, but as "postgres", newly created tables
would not be accessible for the "icingadb" user, resulting in Icinga DB
and Icinga DB Web to fail.

Reported in #987.