history: StreamSorter for Notifications Callback
The StreamSorter was added to history, allowing to collect messages from
multiple Redis streams, sorting them based on the timestamp in the
Stream ID, and ejecting them back.

This is used for the callback stage, required by Icinga Notification. In
the Notification context, an ordered stream is required.

Despite my best intention, it felt like I have created an Erlang.